% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_ind_equations.R
\name{Bass_ind_equations}
\alias{Bass_ind_equations}
\title{Seabass bioenergetic individual model differential equations}
\usage{
Bass_ind_equations(Param, Temp, G, Food, weight)
}
\arguments{
\item{Param}{vector containing all metabolic parameters}

\item{Temp}{water temperature forcing at time t}

\item{G}{food entering the cage at time t}

\item{Food}{food characterization (Proteins, Lipids, Carbohydrates)}

\item{weight}{individual weight at time t}
}
\value{
model output at time t
}
\description{
Seabass bioenergetic individual model differential equations
}

