% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_ind_main.R
\name{Bass_ind_main}
\alias{Bass_ind_main}
\title{Seabass bioenergetic individual model}
\usage{
Bass_ind_main(userpath, forcings)
}
\arguments{
\item{userpath}{the path where forcing are located}

\item{forcings}{a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees] and feeding rate [g/individual x d]}
}
\value{
A list containing model outputs: weight, excreted quantities and quantities to waste, actual and potential ingestion, temperature limitation functions and metabolic rates
}
\description{
Solves the bioenergetic balance for Seabass
}
