% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_ind_skeleton.R
\name{Bass_ind_skeleton}
\alias{Bass_ind_skeleton}
\title{Creates the folders structure for Seabass individual bioenergetic model}
\usage{
Bass_ind_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Seabass individual bioenergetic model
}

