% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_pop_pre.R
\name{Bass_pop_pre}
\alias{Bass_pop_pre}
\title{Seabass bioenergetic population model preprocessor}
\usage{
Bass_pop_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing model forcings}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees] and feeding rate [g/individual x d]
}
\description{
Seabass bioenergetic population model preprocessor
}
