% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_pop_RKsolver.R
\name{Bass_pop_RKsolver}
\alias{Bass_pop_RKsolver}
\title{Solves the Seabass population bioenergetic balance with a 4th order Runge Kutta method}
\usage{
Bass_pop_RKsolver(Param, Temperature, G, Food, IC, times, N)
}
\arguments{
\item{Param}{vector containing all metabolic parameters}

\item{Temperature}{water temperature forcing time series}

\item{G}{food entering the cage at time series}

\item{Food}{food characterization (Proteins, Lipids, Carbohydrates)}

\item{IC}{initial condition on weight}

\item{times}{integration times}

\item{N}{number of individuals time series}
}
\value{
a list containing the fish weight, proteines, lipids and carbohydrates wasted or produced with excretions, potential and actual ingestion rates, temperature limitation functions and metabolic rates
}
\description{
Solves the Seabass population bioenergetic balance with a 4th order Runge Kutta method
}

