% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_pop_skeleton.R
\name{Bass_pop_skeleton}
\alias{Bass_pop_skeleton}
\title{Creates the folders structure for Seabass population model}
\usage{
Bass_pop_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Seabass population model
}

