% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_spatial_pre.R
\name{Bass_spatial_pre}
\alias{Bass_spatial_pre}
\title{Bass bioenergetic spatialized model preprocessor}
\usage{
Bass_spatial_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing forcings used by the model}
}
\value{
a list containing the data used by the main script
}
\description{
Bass bioenergetic spatialized model preprocessor
}
