% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_spatial_pre_int.R
\name{Bass_spatial_pre_int}
\alias{Bass_spatial_pre_int}
\title{Bass bioenergetic spatialized model preprocessor - used inside spatialization loop}
\usage{
Bass_spatial_pre_int(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing forcings used by the model}
}
\value{
a list containing data used by the main script
}
\description{
Bass bioenergetic spatialized model preprocessor - used inside spatialization loop
}
