% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_spatial_skeleton.R
\name{Bass_spatial_skeleton}
\alias{Bass_spatial_skeleton}
\title{Creates the folders structure for Bass spatialized model}
\usage{
Bass_spatial_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Bass spatialized model
}
