% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_ind_main.R
\name{Bream_ind_main}
\alias{Bream_ind_main}
\title{Seabream bioenergetic individual model}
\usage{
Bream_ind_main(userpath, forcings)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{forcings}{a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees] and feeding rate [g/individual x d]}
}
\value{
A list containing model outputs: weight, excreted quantities and quantities to waste, actual and potential ingestion, temperature limitation functions and metabolic rates
}
\description{
Seabream bioenergetic individual model
}
