% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_ind_post.R
\name{Bream_ind_post}
\alias{Bream_ind_post}
\title{Postprocess the Bream indivual bioenergetic model results}
\usage{
Bream_ind_post(userpath, output, times, Dates)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{output}{the output list of the}

\item{times}{the vector containing informations on integration extremes}

\item{Dates}{the vector containing the date}
}
\value{
a list containing the fish weight, proteines, lipids and carbohydrates wasted or produced with excretions, potential and actual ingestion rates, temperature limitation functions and metabolic rates
}
\description{
Postprocess the Bream indivual bioenergetic model results
}

