% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_ind_skeleton.R
\name{Bream_ind_skeleton}
\alias{Bream_ind_skeleton}
\title{Creates the folders structure for Seabream individual bioenergetic model}
\usage{
Bream_ind_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Seabream individual bioenergetic model
}
