% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_pop_equations.R
\name{Bream_pop_equations}
\alias{Bream_pop_equations}
\title{Seabream bioenergetic population model differential equations}
\usage{
Bream_pop_equations(Param, N, Temp, G, Food, weight)
}
\arguments{
\item{Param}{vector containing all metabolic parameters}

\item{N}{the number of individuals at time t}

\item{Temp}{water temperature forcing at time t}

\item{G}{food entering the cage at time t}

\item{Food}{food characterization (Proteins, Lipids, Carbohydrates)}

\item{weight}{individual weight at time t}
}
\value{
model output at time t
}
\description{
Seabream bioenergetic population model differential equations
}
