% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_pop_post.R
\name{Bream_pop_post}
\alias{Bream_pop_post}
\title{Postprocess the Bream population bioenergetic model results}
\usage{
Bream_pop_post(userpath, output, times, Dates, N)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{output}{the output list of the}

\item{times}{the vector containing informations on integration extremes}

\item{Dates}{the vector containing the date}

\item{N}{the number of individuals}
}
\value{
a list containing the fish weight, proteines, lipids and carbohydrates wasted or produced with excretions, potential and actual ingestion rates, temperature limitation functions and metabolic rates
}
\description{
Postprocess the Bream population bioenergetic model results
}

