% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_pop_pre.R
\name{Bream_pop_pre}
\alias{Bream_pop_pre}
\title{Seabream bioenergetic population model preprocessor}
\usage{
Bream_pop_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing model forcings}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees] and feeding rate [g/individual x d]
}
\description{
Preprocesses the data for the bioenergetic balance for Sea Bream
}
