% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_pop_skeleton.R
\name{Bream_pop_skeleton}
\alias{Bream_pop_skeleton}
\title{Creates the folders structure for Seabream population model}
\usage{
Bream_pop_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Seabream population model
}
