% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_dataloader.R
\name{Bream_spatial_dataloader}
\alias{Bream_spatial_dataloader}
\title{Function that loads forcings data for Bream spatialized model and performs the interpolation}
\usage{
Bream_spatial_dataloader(userpath)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees] and feeding rate [g/individual x d]
}
\description{
Function that loads forcings data for Bream spatialized model and performs the interpolation
}

