% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_loop.R
\name{Bream_spatial_loop}
\alias{Bream_spatial_loop}
\title{Bream bioenergetic spatialized model - spatialization loop}
\usage{
Bream_spatial_loop(userpath, forcings)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{forcings}{a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees] and feeding rate [g/individual x d]}
}
\value{
A list containing model outputs: weight, excreted quantities and quantities to waste, actual and potential ingestion, temperature limitation functions and metabolic rates
}
\description{
Solves the bioenergetic balance for Bream
}

