% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_main.R
\name{Bream_spatial_main}
\alias{Bream_spatial_main}
\title{Bream bioenergetic spatialized model - spatialization loop}
\usage{
Bream_spatial_main(userpath, forcings)
}
\arguments{
\item{userpath}{the path where the working folder is located}
}
\value{
saves .nc; .csv and .asc outputs in the 'Outputs' folder
}
\description{
Solves the bioenergetic balance for Bream
}

