% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_post.R
\name{Bream_spatial_post}
\alias{Bream_spatial_post}
\title{Postprocess the Bream spatialized model results}
\usage{
Bream_spatial_post(userpath, output, times, Dates, CS)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{output}{output list containing the output of the RK solver}

\item{times}{the vector containing informations on integration extremes}

\item{Dates}{the vector containing the date}

\item{CS}{the commercial size of Bream}
}
\value{
a list containing the model outputs saved by the main script to .nc; .csv and .asc files
}
\description{
Postprocess the Bream spatialized model results
}
