% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_pre.R
\name{Bream_spatial_pre}
\alias{Bream_spatial_pre}
\title{Bream bioenergetic spatialized model preprocessor}
\usage{
Bream_spatial_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing forcings used by the model}
}
\value{
a list containing the data used in the main script
}
\description{
Bream bioenergetic spatialized model preprocessor
}
