% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_pre_int.R
\name{Bream_spatial_pre_int}
\alias{Bream_spatial_pre_int}
\title{Bream bioenergetic spatialized model preprocessor - used inside spatialization loop}
\usage{
Bream_spatial_pre_int(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing forcings used by the model}
}
\value{
a list containing the data used by the main script
}
\description{
Bream bioenergetic spatialized model preprocessor - used inside spatialization loop
}

