% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bream_spatial_skeleton.R
\name{Bream_spatial_skeleton}
\alias{Bream_spatial_skeleton}
\title{Creates the folders structure for Bream spatialized model}
\usage{
Bream_spatial_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Bream spatialized model
}

