% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clam_ind_equations.R
\name{Clam_ind_equations}
\alias{Clam_ind_equations}
\title{Clam bioenergetic individual model differential equations}
\usage{
Clam_ind_equations(Param, Tint, Phy, DT, POCint, POMint, TSSint, Wd)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{Tint}{the interpolated water temperature at time t}

\item{Phy}{the interpolated phytoplankton at time t}

\item{DT}{the interpolated detritus at time t}

\item{POCint}{the interpolated POC at time t}

\item{POMint}{the interpolated POM at time t}

\item{TSSint}{the interpolated TSS at time t}

\item{Wd}{the weight of the clam at time t}
}
\value{
a list containing the clam weights, temperature limitation functions and metabolic rates at time t
}
\description{
Clam bioenergetic individual model differential equations
}

