% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clam_ind_skeleton.R
\name{Clam_ind_skeleton}
\alias{Clam_ind_skeleton}
\title{Creates the folders structure for Clam individual bioenergetic model}
\usage{
Clam_ind_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Clam individual bioenergetic model
}
