% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clam_pop_dataloader.R
\name{Clam_pop_dataloader}
\alias{Clam_pop_dataloader}
\title{Function that loads forcings data for Clam population model and performs the interpolation}
\usage{
Clam_pop_dataloader(userpath)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees], Chlorophyll a concentration [mgChl-a/m^3], particulated organic carbon (POC) concentration [mgC/l], particulated organic matter (POM) concentration [mgC/l], total suspended solids (TSS) concentration [mg/l]
}
\description{
Function that loads forcings data for Clam population model and performs the interpolation
}
