% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clam_pop_main.R
\name{Clam_pop_main}
\alias{Clam_pop_main}
\title{Clam bioenergetic population model}
\usage{
Clam_pop_main(userpath, forcings)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{forcings}{a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees], Chlorophyll a concentration [mgChl-a/m^3], particulated organic carbon (POC) concentration [mgC/l], particulated organic matter (POM) concentration [mgC/l], total suspended solids (TSS) concentration [mg/l]}
}
\value{
A list containing model outputs: weights, temperature limitation functions and metabolic rates
}
\description{
Clam bioenergetic population model
}
