% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clam_pop_pre.R
\name{Clam_pop_pre}
\alias{Clam_pop_pre}
\title{Clam bioenergetic population model preprocessor}
\usage{
Clam_pop_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing model forcings}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees], Chlorophyll a concentration [mgChl-a/m^3], particulated organic carbon (POC) concentration [mgC/l], particulated organic matter (POM) concentration [mgC/l], total suspended solids (TSS) concentration [mg/l]
}
\description{
Clam bioenergetic population model preprocessor
}

