% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clam_pop_skeleton.R
\name{Clam_pop_skeleton}
\alias{Clam_pop_skeleton}
\title{Creates the folders structure for Clam population model}
\usage{
Clam_pop_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Clam population model
}
