% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_ind_RKsolver.R
\name{ClamF_ind_RKsolver}
\alias{ClamF_ind_RKsolver}
\title{Solves the Clam bioenergetic balance (alternative version) with a 4th order Runge Kutta method}
\usage{
ClamF_ind_RKsolver(Param, times, IC, Tint, Chlint)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{times}{integration extremes and integration timestep}

\item{IC}{initial condition}

\item{Tint}{the interpolated water temperature time series}

\item{Chlint}{the interpolated chlorophyll a time series}
}
\value{
a list containing the clam weights, temperature limitation functions and metabolic rates
}
\description{
Solves the Clam bioenergetic balance (alternative version) with a 4th order Runge Kutta method
}
