% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_ind_skeleton.R
\name{ClamF_ind_skeleton}
\alias{ClamF_ind_skeleton}
\title{Creates the folders structure for Clam individual bioenergetic model (alternative version)}
\usage{
ClamF_ind_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Clam individual bioenergetic model (alternative version)
}
