% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_pop_equations.R
\name{ClamF_pop_equations}
\alias{ClamF_pop_equations}
\title{Clam bioenergetic population model differential equations (alternative version)}
\usage{
ClamF_pop_equations(Param, Tint, Chlint, Ww)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{Tint}{the interpolated water temperature at time t}

\item{Chlint}{the interpolated chlorophyll at time t}

\item{Ww}{clam wet weight at time t}
}
\value{
a list containing the clam weights, temperature limitation functions and metabolic rates at time t
}
\description{
Clam bioenergetic population model differential equations (alternative version)
}
