% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_pop_loop.R
\name{ClamF_pop_loop}
\alias{ClamF_pop_loop}
\title{Function that runs the Monte Carlo simulation for the Clam population model (alternative version)}
\usage{
ClamF_pop_loop(Param, times, IC, Tint, Chlint, N, userpath)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{times}{integration extremes and integration timestep}

\item{IC}{initial condition}

\item{Tint}{the interpolated water temperature time series}

\item{Chlint}{the interpolated chlorophyll a time series}

\item{N}{time series with number of individuals}

\item{userpath}{the path where the working folder is located}
}
\value{
a list with RK solver outputs
}
\description{
Function that runs the Monte Carlo simulation for the Clam population model (alternative version)
}

