% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_pop_pre.R
\name{ClamF_pop_pre}
\alias{ClamF_pop_pre}
\title{Clam bioenergetic population model (alternative version) preprocessor}
\usage{
ClamF_pop_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing model forcings}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees], Chlorophyll a concentration [mgChl-a/m^3]
}
\description{
Clam bioenergetic population model (alternative version) preprocessor
}

