% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_pop_skeleton.R
\name{ClamF_pop_skeleton}
\alias{ClamF_pop_skeleton}
\title{Creates the folders structure for Clam population model (alternative version)}
\usage{
ClamF_pop_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Clam population model (alternative version)
}

