% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_ind_equations.R
\name{Mussel_ind_equations}
\alias{Mussel_ind_equations}
\title{Mussel bioenergetic individual model differential equations}
\usage{
Mussel_ind_equations(Param, Tint, Phyint, DTint, POCint, Ccont, Ncont, Pcont,
  POMint, TSSint, Wb, R, t, trip)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{Tint}{the interpolated water temperature at time t}

\item{Phyint}{the interpolated phytoplankton at time t}

\item{DTint}{the interpolated detritus at time t}

\item{POCint}{the interpolated POC at time t}

\item{Ccont}{the C/C content of the POC at time t}

\item{Ncont}{the N/C content of POC at time t}

\item{Pcont}{the P/C content of POC at time t}

\item{POMint}{the interpolated POM at time t}

\item{TSSint}{the interpolated TSS at time t}

\item{Wb}{the somatic tissue dry weight at time t}

\item{R}{the gondadic tissue dry weight at time t}

\item{t}{the time}

\item{trip}{vector containing the flags with resting periods}
}
\value{
the outputs at time t
}
\description{
Mussel bioenergetic individual model differential equations
}

