% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_ind_skeleton.R
\name{Mussel_ind_skeleton}
\alias{Mussel_ind_skeleton}
\title{Creates the folders structure for Mussel individual bioenergetic model}
\usage{
Mussel_ind_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Mussel individual bioenergetic model
}

