% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_pop_equations.R
\name{Mussel_pop_equations}
\alias{Mussel_pop_equations}
\title{Mussel bioenergetic population model differential equations}
\usage{
Mussel_pop_equations(
  Param,
  N,
  Tint,
  Phyint,
  DTint,
  POCint,
  Ccont,
  Ncont,
  Pcont,
  POMint,
  TSSint,
  Wb,
  R,
  t,
  trip
)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{N}{the number of individuals at time t}

\item{Tint}{the interpolated water temperature at time t}

\item{Phyint}{the interpolated phytoplankton at time t}

\item{DTint}{the interpolated detritus at time t}

\item{POCint}{the interpolated POC at time t}

\item{Ccont}{the C/C content of the POC at time t}

\item{Ncont}{the N/C content of POC at time t}

\item{Pcont}{the P/C content of POC at time t}

\item{POMint}{the interpolated POM at time t}

\item{TSSint}{the interpolated TSS at time t}

\item{Wb}{the somatic tissue dry weight at time t}

\item{R}{the gondadic tissue dry weight at time t}

\item{t}{the time}

\item{trip}{vector containing the flags with resting periods}
}
\value{
the outputs at time t
}
\description{
Mussel bioenergetic population model differential equations
}
