% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_pop_loop.R
\name{Mussel_pop_loop}
\alias{Mussel_pop_loop}
\title{Function that runs the Monte Carlo simulation for the Mussel population model}
\usage{
Mussel_pop_loop(
  Param,
  times,
  IC,
  Tint,
  Phyint,
  DTint,
  POCint,
  Ccont,
  Ncont,
  Pcont,
  POMint,
  TSSint,
  N,
  userpath
)
}
\arguments{
\item{Param}{a vector containing model parameters}

\item{times}{integration extremes and integration timestep}

\item{IC}{initial condition}

\item{Tint}{the interpolated water temperature time series}

\item{Phyint}{the interpolated phytoplankton time series}

\item{DTint}{the interpolated detritus time series}

\item{POCint}{the interpolated POC time series}

\item{Ccont}{the C/C content of the POC}

\item{Ncont}{the N/C content of POC}

\item{Pcont}{the P/C content of POC}

\item{POMint}{the interpolated POM time series}

\item{TSSint}{the interpolated TSS time series}

\item{N}{time series with number of individuals}

\item{userpath}{the path where the working folder is located}
}
\value{
a list with RK solver outputs
}
\description{
Function that runs the Monte Carlo simulation for the Mussel population model
}
