% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_pop_post.R
\name{Mussel_pop_post}
\alias{Mussel_pop_post}
\title{Postprocess the Mussel population bioenergetic model results}
\usage{
Mussel_pop_post(userpath, output, times, Dates, N)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{output}{the output list of the}

\item{times}{the vector containing informations on integration extremes}

\item{Dates}{the vector containing the date}

\item{N}{the number of individuals}
}
\value{
a list containing the weights of the mussel, the excreted CNP, the mussel CNP, temperature limitation functions, metabolic rates, oxygen consumption
}
\description{
Postprocess the Mussel population bioenergetic model results
}

