% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_pop_pre.R
\name{Mussel_pop_pre}
\alias{Mussel_pop_pre}
\title{Mussel bioenergetic population model preprocessor}
\usage{
Mussel_pop_pre(userpath, forcings)
}
\arguments{
\item{userpath}{the path where folder containing model inputs and outputs is located}

\item{forcings}{a list containing forcings used by the model}
}
\value{
a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees], Chlorophyll a concentration [mgChl-a/m^3], particulated organic carbon (POC) concentration [mgC/l] and its characterization in terms of C/P and N/P molar ratios, particulated organic matter (POM) concentration [mgC/l], total suspended solids (TSS) concentration [mg/l]
}
\description{
Mussel bioenergetic population model preprocessor
}

