% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_pop_skeleton.R
\name{Mussel_pop_skeleton}
\alias{Mussel_pop_skeleton}
\title{Creates the folders structure for Mussel population model}
\usage{
Mussel_pop_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Mussel population model
}
