% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mussel_spatial_skeleton.R
\name{Mussel_spatial_skeleton}
\alias{Mussel_spatial_skeleton}
\title{Creates the folders structure for Mussel spatialized model}
\usage{
Mussel_spatial_skeleton(userpath)
}
\arguments{
\item{userpath}{the path where forcing are located}
}
\description{
Creates the folders structure for Mussel spatialized model
}

