% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_fun.R
\name{Pop_fun}
\alias{Pop_fun}
\title{Function that solves the population dynamics equations including discontinuities}
\usage{
Pop_fun(Nseed, mort, manag, times)
}
\arguments{
\item{Nseed}{number of seeded individuals}

\item{mort}{mortality rate}

\item{manag}{list of management actions (seeded/harvested individuals)}

\item{times}{vector containing informations on integration times}
}
\value{
a time series with the number of individuals
}
\description{
Function that solves the population dynamics equations including discontinuities
}

