% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bass_spatial_loop.R
\name{Bass_spatial_loop}
\alias{Bass_spatial_loop}
\title{Bass bioenergetic spatialized model - spatialization loop}
\usage{
Bass_spatial_loop(userpath, forcings)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{forcings}{a list containing the time series in the odd positions and realted forcings in the even positions. Forcings returned are: Water temperature [Celsius degrees], Chlorophyll a concentration [mgChl-a/m^3], particulated organic carbon (POC) concentration [mgC/l] and its characterization in terms of C/P and N/P molar ratios, particulated organic matter (POM) concentration [mgC/l], total suspended solids (TSS) concentration [mg/l]}
}
\value{
a list containing the outputs that main script saves to .nc; .csv and .asc files
}
\description{
Solves the bioenergetic balance for Bass
}
