% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamF_pop_post.R
\name{ClamF_pop_post}
\alias{ClamF_pop_post}
\title{Postprocess the Clam population bioenergetic model results (alternative model)}
\usage{
ClamF_pop_post(userpath, output, times, Dates, N, CS)
}
\arguments{
\item{userpath}{the path where the working folder is located}

\item{output}{output list containing the output of the RK solver}

\item{times}{the vector containing informations on integration extremes}

\item{Dates}{the vector containing the date}

\item{N}{the number of individuals}

\item{CS}{the commercial size of Clam}
}
\value{
a list containing the clam weights, temperature limitation functions and metabolic rates
}
\description{
Postprocess the Clam population bioenergetic model results (alternative model)
}
